# backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Realtime Chat API
  version: 1.0.0
  description: API for multi-device auth and chat system
servers:
  - url: http://localhost:3000/api/v1
paths:
  /auth/sign-up:
    post:
      tags:
        - Auth
      summary: Sign up a new user
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignUpResponse"
        '400':
          description: Invalid input

  /auth/sign-in:
    post:
      tags:
        - Auth
      summary: Sign in and issue tokens
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignInResponse'
        '401':
          description: Invalid credentials
  
  /auth/sign-out:
    post:
      tags:
        - Auth
      summary: Sign out current device
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signed out successfully
  
  /auth/sign-out-all:
    post:
      tags:
        - Auth
      summary: Sign out all devices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signed out from all devices

  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token using HttpOnly cookie
      description: Refreshes the access token using the HttpOnly refresh token cookie; no request body is needed
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/SignInResponse"

  /users/me:
    get:
      tags:
        - User
      summary: Get current user info
      description: Returns the currently authenticated user's profile. Requires Bearer access token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized, invalid or expired token

  /chats:
    post:
      tags:
        - Chat
      summary: Create a new chat (group or private)
      description: Creates a new chat based on the isGroup flag, with an optional name for group chats.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"
      responses:
        '200':
          description: Chat created successfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/CreateChatResponse"
        '401':
          description: Invalid credentials

    get:
      tags:
        - Chat
      summary: Get chat list
      description: Retrieve a paginated list of chat messages with optional search and cursor-based pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter chat by user name or chat name
        - name: take
          in: query
          required: false
          schema:
            type: integer
            example: 20
          description: Number of chats to retrieve
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: date-time
            example: "2026-04-14T10:00:00.000Z"
          description: Cursor for pagination (fetch chats before this timestamp)
      responses:
          '200':
            description: Successfully retrieved chat list
            content: 
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateChatResponse"
          '401':
            description: Invalid credentials
          '404':
            description: Chat or user not found

  /chats/{chatId}/users:
    post:
      tags:
        - Chat
      summary: Add user to chat
      description: Adds a user to the specified chat. The user must not already exist in the chat.
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the chat
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/AddUsersToChatRequest"
      responses:
        '200':
          description: User added to chat successfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/AddUsersToChatResponse"
        '400':
          description: User already exists in the chat or invalid input
        '401':
          description: Invalid credentials
        '404':
          description: Chat or user not found
      
  /chats/{chatId}/messages:
    post:
      tags:
        - Chat
      summary: Send message to chat
      description: Sends a message to the specified chat. The authenticated user must be a member of the chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
            description: Unique identifier of the chat
      requestBody: 
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/SendMessageResponse"
        '400':
          description: Invalid request payload
        '401':
          description: Invalid credentials
        '403':
          description: Forbidden (user is not a member of the chat)
        '404':
          description: Chat not found
    
    get:
      tags:
        - Chat
      summary: Get messages of a chat
      description: Retrieve a paginated list of messages in a specific chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
          description: ID of the chat
        - name: take
          in: query
          required: false
          schema:
            type: integer
            example: 20
          description: Number of messages to retrieve
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            format: date-time
            example: "2026-04-14T10:00:00.000Z"
          description: Cursor for pagination (fetch messages before this timestamp)
        - name: after
          in: query
          required: false
          schema: 
            type: string
            format: date-time
            example: "2026-04-14T10:00:00.000Z"
          description: Fetch messages after this timestamp
        - name: messageId
          in: query
          required: false
          schema: 
            type: string
          description: "Unique identifier of the message"
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SendMessageResponse"
        '400':
          description: Invalid request payload
        '401':
          description: Invalid credentials
        '403':
          description: Forbidden (user is not a member of the chat)
        '404':
          description: Chat not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignUpRequest:
      type: object
      required:
        - username
        - email
        - password
        - fullName
      properties:
        username:
          type: string
          description: The user's username, must be unique
          example: admin
        email: 
          type: string
          format: email
          example: admin@gmail.com
          description: The user's email address, must be unique
        password:
          type: string
          format: password
          minLength: 8
          pattern: "^(?=.*[A-Z])(?=.*[!@#$%^&*]).*$"
          example: Admin@123456
          description: The user's password, minimum 8 characters recommended
        fullName:
          type: string
          example: 'Admin Realtime Chat'
          description: Full name of the user

    SignUpResponse:
      type: object
      properties:
        id: 
          type: string
          description: Unique identifier of the user
        username:
          type: string
          description: Username of the user
        email:
          type: string
          description: Email of the user
        fullName:
          type: string
          description: Full name of the user
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user account was created
        
    SignInRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
          description: Registered username of the user
        password:
          type: string
          format: password
          example: Admin@123456
          description: Password for authentication

    SignInResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token used for calling protected APIs
        tokenType:
          type: string
          description: Type of token, usually "Bearer"
          example: Bearer
        expiresAt:
          type: string
          format: date-time
          description: Expiration time of the access token

    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the user
        username:
          type: string
          description: Username of the user
        email:
          type: string
          description: Email of the user
        fullName:
          type: string
          description: Full name of the user
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user account was created

    CreateChatRequest:
      type: object
      required:
        - users
      properties:
        name:
          type: string
          description: Name of the chat (required for group chats, optional for private chats)
        users: 
          type: array
          description: List of user IDs to be added to the chat
          items:
            type: string
          example: ["user_1", "user_2"]
        
    CreateChatResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the created chat
        name: 
          type: string
          description: Name of the chat
        isGroup:
          type: boolean
          description: Indicates whether the chat is a group chat
        createdAt:
          type: string
          format: date-time
          description: The date and time when the chat was created

    AddUsersToChatRequest:
      type: object
      required:
        - userIds
      properties:
        userIds:
          type: array          
          items:
            type: string
            description: Unique identifier of a user
          description: List of unique user IDs to be added to the chat

    AddUsersToChatResponse:
      type: object
      properties:
        chatId:
            type: string
            description: Unique identifier of the chat
        users: 
          type: array
          description: List of users added to the chat
          items: 
            type: object
            properties:
              id: 
                type: string
                description: Unique identifier of the ChatUser record
              userId: 
                type: string
                description: Unique identifier of the user added to the chat
    
    SendMessageRequest:
      type: object
      required:
        - content
        - tempId
      properties:
        content:
          type: string
          example: Hello World
          description: The textual content of the message to be sent
        tempId:
          type: string
          description: Temporary client-side ID used for message tracking before server persistence
    
    SendMessageResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the message
        senderId:
          type: string
          description: Unique identifier of the user who sent the message
        chatId: 
          type: string
          description: Unique identifier of the chat where the message was sent
        tempId:
          type: string
          description: Temporary client-generated identifier used to track and reconcile messages before the server assigns a permanent id
        content: 
          type: string
          description: The textual content of the message
        createdAt:
          type: string
          format: date-time
          description: The date and time when the message was created
        isMe:
          type: boolean
          description:  Indicates whether the message was sent by the current authenticated user